Setup webpack for development and builds
webpack with the elm-webpack-loader and the webpack-dev-server make it
very easy to compile an elm app and automatically compile it while
developing. The configuration is inspired by the example of
elm-webpack-loader.
See: https://github.com/rtfeldman/elm-webpack-loader

I added the -p option to the build script to make sure the files are
minified when the application is build.
See: https://github.com/webpack/docs/wiki/cli#production-shortcut--p

I also chose to compress the served files for transfering smaller file
sizes while developing.