Setup a very simple app with navigation
The structure of this app follows a mix of the article "How I Structure
Web Apps", an example repo from that article and an example of the hop
package.
See:
http://blog.jenkster.com/2016/04/how-i-structure-elm-apps.html
https://github.com/krisajenkins/elm-dialog/tree/master/examples/Advanced
https://github.com/sporto/hop/tree/master/examples/full

The idea to split your app into App, State, Types, View, Routing seems
like a good idea.

I chose hop as the routing framework as it seems to be the most popular
with lots of documentation.